dtoverlays: adds the definitions for the HiFiBerry 8-channel ADC
Additions and changes for the 8 channel ADC card. This card uses only
HW-controlled devices which allows the uses of the 'dummy-dai'.
It will run only on a PI5 as it requires the designware I2S0 module.

The necessary output lanes I2S0_DI[0..3] are claimed from within the
DT overlay.

Signed-off-by: j-schambacher <joerg@hifiberry.com>