dts: rp1: Disable DMA usage for UART0
Some recent DMA changes have led to data loss in UART0 on Pi 5. It also
seems that even prior to these changes there was a problem with aborted
transfers.

As this is the only RP1 UART configured for DMA, it is better to remove
the DMA usage until it is shown to be reliable.

Link: https://github.com/raspberrypi/linux/issues/6365

Signed-off-by: Phil Elwell <phil@raspberrypi.com>